refactor: move package and update doc
