+++ build.gradle.kts
@@ -11,7 +11,7 @@ java {
 dependencies {
 }
 
-group = "com.github.foodiestudio.plugin"
+group = "com.github.foodiestudio"
 version = "0.0.1-SNAPSHOT"
 
 // 为 buildSrc 里的 Plugin 创建对应的 id，这样才能在 plugins{} 中使用
@@ -19,7 +19,7 @@ gradlePlugin {
     plugins {
         create("rustDesktop") {
             id = "rust-desktop"
-            implementationClass = "com.github.foodiestudio.plugin.CargoDesktopPlugin"
+            implementationClass = "com.github.foodiestudio.plugin.rust.CargoDesktopPlugin"
         }
     }
 }
