+++ src/main/java/com/github/foodiestudio/plugin/rust/CargoDesktopPlugin.kt
@@ -1,8 +1,7 @@
-package com.github.foodiestudio.plugin
+package com.github.foodiestudio.plugin.rust
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.kotlin.dsl.register
 
 class CargoDesktopPlugin : Plugin<Project> {
